refactored training modules to use same terminology
