+++ src/App.js
@@ -10,14 +10,14 @@ import SignUp from './components/signUpForm/SignUpOverlay/SignUpOverlay'
 import AdminDash from './components/AdminDash/AdminDashboard'
 import Profile from './components/AdminDash/Profile/Profile'
 import Application from './components/signUpForm/Application/Application'
-import Training from './components/Training/Module3/TrainingOverlay'
+import Training from './components/Training/module-3/TrainingOverlay'
 import ModuleLetter from './components/Training/module-0/trainingLetter'
-import StartTraining from './components/Training/module-1/moduleOneLetter'
-import ModuleOne from './components/Training/module-1/moduleOne'
+import StartTraining from './components/Training/module-1/module1-1Letter'
+import ModuleOne from './components/Training/module-1/module1-1'
 import ModuleTwo from './components/Training/module-2'
-import ModuleThree from './components/Training/Module3/TrainingOverlay'
+import ModuleThree from './components/Training/module-3/TrainingOverlay'
 import ModuleFour from './components/Training/module-4'
-import ModuleFive from './components/Training/module-5/moduleFive'
+import ModuleFive from './components/Training/module-5/module5-1'
 import UserProfile from './components/AdminDash/Profile/User/UserProfile'
 
 function App() {
