+++ src/components/Training/module-1/module1-1.test.js
@@ -1,14 +1,14 @@
 import React from 'react'
 import { renderWithReduxAndRouter as render } from '../../../utils/testHelpers'
 import ModuleOne from './moduleOne'
-import StartTraining from './moduleOneLetter'
+import StartTraining from './module1-1Letter'
 import '@testing-library/jest-dom/extend-expect'
 
 test('ModuleOne Renders', async () => {
   render(<ModuleOne />)
 })
 
-test('ModuleOneLetter Renders', async () => {
+test('Module1-1Letter Renders', async () => {
   render(<StartTraining />)
 })
