+++ src/components/Training/module-2/components/module2-2.js
@@ -45,7 +45,7 @@ const simpleStepData = [
   },
 ]
 
-export const BestPractices = ({ handleBack, errors, touched }) => {
+export const Module22 = ({ handleBack, errors, touched }) => {
   return (
     <Wrapper>
       <Title>
