+++ src/components/Training/module-2/components/module2-3.js
@@ -178,7 +178,7 @@ const BeyondTheMeal = () => {
   )
 }
 
-export const MealTips = ({ handleBack, ...props }) => {
+export const Module23 = ({ handleBack, ...props }) => {
   return (
     <Wrapper>
       <Title>
