+++ src/components/Training/module-3/TrainStyle.js
