+++ src/components/Training/module-3/module3-4.js
@@ -16,7 +16,7 @@ import {
   Submit,
 } from '../TrainStyles'
 
-export default function TrainingThreeFour() {
+export default function Module34() {
   const { push } = useHistory()
 
   const handleSubmit = () => {
